Remove unused files and refactor PHP scripts to enable CORS for API endpoints

+++ note_app/delete.php
@@ -1,4 +1,14 @@
 <?php
+// Enable CORS
+header('Access-Control-Allow-Origin: *');
+header('Access-Control-Allow-Methods: GET, POST, OPTIONS');
+header('Access-Control-Allow-Headers: Content-Type');
+header('Content-Type: application/json');
+
+if ($_SERVER['REQUEST_METHOD'] === 'OPTIONS') {
+    exit(0);
+}
+
 $connection = new mysqli("mysql", "root", "toor", "coba");
 $id = $_POST['id'];
 $result = mysqli_query($connection, "delete from note_app where
